added a couple tests, separated features into its own component folder, gitignore snapshots
