+++ .gitignore
@@ -2,3 +2,4 @@ node_modules
 .config.js
 bundle.js
 coverage
+__snapshots__
